dhwydhjs

+++ modules/WXRegisterIII/WXRegisterIII.py
@@ -76,22 +76,20 @@ class WeiXinRegister:
             return "again"
 
         z.input( PhoneNumber )
-        d( text='用短信验证码登录' ).click( )
-        z.sleep( 1.5 )
         if d( text='下一步' ).exists:
             d( text='下一步' ).click( )
             z.sleep( 1.5 )
 
+        d( text='用短信验证码登录' ).click( )
+        z.sleep(1.5)
+
+        d( text='获取验证码' ).click( )
+        z.sleep( 1 )
+
         if d( text='确认手机号码' ).exists:
             d( text='确定' ).click( )
-            z.sleep( 1.5 )
-            if d( text='确认手机号码' ).exists:
-                d( text='确定' ).click( )
 
         z.sleep( 3 )
-        if d( textContains='正在验证' ).exists:
-            z.sleep( 35 )
-            z.heartbeat( )
         try:
             code = self.scode.GetVertifyCode( PhoneNumber, self.scode.WECHAT_REGISTER )  # 获取接码验证码
             self.scode.defriendPhoneNumber( PhoneNumber, self.scode.WECHAT_REGISTER )
@@ -103,18 +101,11 @@ class WeiXinRegister:
             z.toast( PhoneNumber + '手机号,获取不到验证码' )
             return "again"
 
-        z.heartbeat( )
-        d( text='请输入验证码' ).click( )
-        d( text='请输入验证码' ).click( )
         z.input( code )
-
         print( '手机号码：' + PhoneNumber + '验证码：' + code )
-        x_02 = int( numbers[0]['x02'] ) + 1
-        para = {"phoneNumber": PhoneNumber, 'x_02': x_02, 'x_19': 'WXRegister'}
-        self.repo.PostInformation( saveCate, para )
 
         z.sleep( 2 )
-        d( text='下一步', className='android.widget.Button' ).click( )
+        d( text='登录', className='android.widget.Button' ).click( )
 
         z.sleep( 5 )
         z.heartbeat( )
@@ -125,14 +116,14 @@ class WeiXinRegister:
             if d( textContains='如果这不是你本人操作，你的短信内容已经泄露。请检查手机是否被植入木马导致短信被转发。' ).exists:
                 d( text='确定' ).click( )
 
+            featureCodeInfo = z.get_serial( "com.tencent.mm" )
             # 信息入库
-            para = {"phoneNumber": PhoneNumber, 'x_05': 'YES', 'x_19': 'WXRegister',
-                    'x_20': str( slotnum ) + '_' + d.server.adb.device_serial( ), 'x_26': '登陆状态YSE'}
+            para = {"phoneNumber": PhoneNumber, 'x_02': slotnum, 'x_03': d.server.adb.device_serial(),
+                    'x_04': featureCodeInfo, 'x_05': nowTime, 'x_19': 'WXRegister'}
             self.repo.PostInformation( saveCate, para )
 
             # 微信号入卡槽
-            featureCodeInfo = z.get_serial( "com.tencent.mm" )
-            self.slot.backup( slotnum, str( slotnum ) + '_' + PhoneNumber + "_" + featureCodeInfo )  # 卡槽号，手机号
+            self.slot.backup( slotnum, str( slotnum ) + '_' + PhoneNumber )  # 卡槽号，手机号
             return "success"
 
         if d( textContains='当前手机号一个月内已成功注册微信号' ).exists:
@@ -153,75 +144,45 @@ class WeiXinRegister:
                 nickname = ''.join( [i for i in nicknameList] )
                 z.input(nickname)
                 d(text='注册').click()
+                z.sleep(2)
+
+                if d(textContains='相同手机号不可频繁重复注册微信帐号。').exists:
+                    self.repo.DeleteInformation( saveCate, PhoneNumber )
+                    return "again"
+
                 while not d(text='查找你的微信朋友').exists:
                     d.dump( compressed=False )
                     z.sleep( 3 )
 
-                if d(text='好').exists:
-                    d(text='好').click()
-                    z.sleep(8)
-
-                if d( text='发现' ) and d( text='我' ) and d( text='通讯录' ).exists:
-                    # 信息入库
-                    para = {"phoneNumber": PhoneNumber, 'x_05': 'YES', 'x_19': 'WXRegister',
-                            'x_20': str( slotnum ) + '_' + d.server.adb.device_serial( ), 'x_26': '登陆状态YSE'}
-                    self.repo.PostInformation( saveCate, para )
-
-                    # 微信号入卡槽
-                    featureCodeInfo = z.get_serial( "com.tencent.mm" )
-                    self.slot.backup( slotnum, str( slotnum ) + '_' + PhoneNumber + "_" + featureCodeInfo )  # 卡槽号，手机号
-                    return "success"
-
-
-        n = 0
-        while True:
-            z.sleep(8)
-            n = n + 1
-            if d(text='下一步').exists:
-                d(text='下一步').click()
-                z.sleep(1.5)
-
-            if d( text='确认手机号码' ).exists:
-                d( text='确定' ).click( )
-                z.sleep( 1.5 )
-                if d( text='确认手机号码' ).exists:
-                    d( text='确定' ).click( )
-            z.sleep( 3 )
-
-            if d( textContains='看看手机通讯录' ).exists:
-                d( text='是' ).click( )
-                z.sleep( 15 )
-                z.heartbeat( )
-                if d( textContains='如果这不是你本人操作，你的短信内容已经泄露。请检查手机是否被植入木马导致短信被转发。' ).exists:
-                    d( text='确定' ).click( )
-
+                featureCodeInfo = z.get_serial( "com.tencent.mm" )
                 # 信息入库
-                number_info = self.repo.GetInformation( saveCate, PhoneNumber )
-                if number_info[0]['x05'] is None:
-                    para = {"phoneNumber": PhoneNumber, 'x_05': 'YES', 'x_19': 'WXRegister', 'x_26': '登陆状态YSE',
-                            'x_20': str( slotnum ) + '_' + d.server.adb.device_serial( )}
-                else:
-                    para = {"phoneNumber": PhoneNumber, 'x_05': number_info[0]['x05'], 'x_19': 'WXRegister',
-                            'x_26': '登陆状态YSE',
-                            'x_20': str( slotnum ) + '_' + d.server.adb.device_serial( )}
+                para = {"phoneNumber": PhoneNumber, 'x_02': slotnum, 'x_03': d.server.adb.device_serial( ),
+                        'x_04': featureCodeInfo, 'x_05': nowTime, 'x_19': 'WXRegister'}
                 self.repo.PostInformation( saveCate, para )
 
                 # 微信号入卡槽
-                featureCodeInfo = z.get_serial( "com.tencent.mm" )
-                self.slot.backup( slotnum, str( slotnum ) + '_' + PhoneNumber + "_" + featureCodeInfo )  # 卡槽号，手机号
+                self.slot.backup( slotnum, str( slotnum ) + '_' + PhoneNumber )  # 卡槽号，手机号
+
+                if d(text='好').exists:
+                    d(text='好').click()
+                    z.sleep(10)
+
                 return "success"
-            elif d( textContains='是否立即验证' ).exists:
+
+        n = 0
+        while True:
+            z.sleep(5)
+            n = n + 1
+            if d( text='登录', className='android.widget.Button' ).exists:
+                d( text='登录', className='android.widget.Button' ).click( )
+                z.sleep( 8 )
+
+            if d( textContains='是否立即验证' ).exists:
                 d( text='确定' ).click( )
-                z.sleep( 1.5 )
-                if d( textContains='是否立即验证' ).exists:
-                    d( text='确定' ).click( )
-                z.sleep( 3 )
+                z.sleep(8)
             elif d( text='声纹验证' ).exists:
                 d( className='android.widget.ImageView', description='返回' ).click( )
-                x_27 = int( numbers[0]['x27'] ) + 1
-                para = {"phoneNumber": PhoneNumber, 'x_01': "exist", 'x_03': nowTime,
-                        'x_19': 'WXRegister', 'x_27': x_27}
-                self.repo.PostInformation( saveCate, para )
+                self.repo.DeleteInformation( saveCate, PhoneNumber )
                 break
             elif d( text='系统检测到帐号有被盗风险。为了你的帐号安全，新设备的登录请求会被拒绝。请使用常用设备登录微信。' ).exists:
                 d( text='确定' ).click( )
@@ -233,16 +194,12 @@ class WeiXinRegister:
                 self.repo.DeleteInformation( saveCate, PhoneNumber )
                 break
 
-            z.heartbeat( )
-            while d( text='确认登录' ).exists:
+            if d( text='确认登录' ).exists:
                 if n == 1:
                     d( className='android.widget.ImageView', description='返回' ).click( )
                     continue
                 elif n == 2:
                     self.repo.DeleteInformation( saveCate, PhoneNumber )
-                else:
-                    para = {"phoneNumber": PhoneNumber, 'x_01': "exist", 'x_03': nowTime, 'x_19': 'WXRegister'}
-                    self.repo.PostInformation( saveCate, para )
                 break
 
             if d( text='设置密码' ).exists:
@@ -269,39 +226,41 @@ class WeiXinRegister:
 
                 if d( textContains='如果这不是你本人操作，你的短信内容已经泄露。请检查手机是否被植入木马导致短信被转发。' ).exists:
                     d( text='确定' ).click( )
-
-                para = {"phoneNumber": PhoneNumber, 'x_05': 'YES', 'x_19': 'WXRegister',
-                        'x_20': str( slotnum ) + '_' + d.server.adb.device_serial( ), 'x_21': password,
-                        'x_26': '登陆状态YSE'}
-                self.repo.PostInformation( saveCate, para )
                 featureCodeInfo = z.get_serial( "com.tencent.mm" )
-                self.slot.backup( slotnum, str( slotnum ) + '_' + PhoneNumber + "_" + featureCodeInfo )  # 卡槽号，手机号
+                # 入库信息
+                para = {"phoneNumber": PhoneNumber, 'x_02': slotnum, 'x_03': d.server.adb.device_serial(), 'x_04':featureCodeInfo,'x_05':nowTime,'x_06':password, 'x_19': 'WXRegister'}
+                self.repo.PostInformation( saveCate, para )
+
+                # 入卡槽信息
+                self.slot.backup( slotnum, str( slotnum ) + '_' + PhoneNumber )  # 卡槽号，手机号
                 return "success"
 
-            x_value = ''
             if d( text='验证身份' ).exists:
                 z.sleep( 3 )
                 if d( description='通过扫码验证身份', className='android.view.View', index=1 ).exists:
-                    para = {"phoneNumber": PhoneNumber, 'x_01': "exist", 'x_03': nowTime, 'x_19': 'WXRegister'}
+                    featureCodeInfo = z.get_serial( "com.tencent.mm" )
+                    # 入库信息
+                    para = {"phoneNumber": PhoneNumber, 'x_02': slotnum, 'x_03': d.server.adb.device_serial( ),'x_04': featureCodeInfo, 'x_05': nowTime, 'x_19': 'WXRegister'}
                     self.repo.PostInformation( saveCate, para )
+
+                    # 入卡槽信息
+                    self.slot.backup( slotnum, str( slotnum ) + '_' + PhoneNumber )  # 卡槽号，手机号
                     break
 
                 if d( descriptionContains='验证失败', className='android.view.View' ).exists:
                     d( descriptionContains='关闭页面', className='android.view.View', index=3 ).click( )
-                    if x_value != '':
-                        para = {"phoneNumber": PhoneNumber, 'x_01': "exist", "x_key": x_value,
-                                'x_19': 'WXRegister'}
-                    self.repo.PostInformation( saveCate, para )
                     break
 
                 if d( descriptionContains='验证通过', className='android.view.View' ).exists:
                     d( descriptionContains='关闭页面', className='android.view.View', index=3 ).click( )
-                    if x_value == '':
-                        x_value = 'YSE'
-                    para = {"phoneNumber": PhoneNumber, 'x_01': "exist", 'x_05': x_value, 'x_19': 'WXRegister'}
-                    self.repo.PostInformation( saveCate, para )
                     featureCodeInfo = z.get_serial( "com.tencent.mm" )
-                    self.slot.backup( slotnum, str( slotnum ) + '_' + PhoneNumber + "_" + featureCodeInfo )  # 卡槽号，手机号
+                    # 入库信息
+                    para = {"phoneNumber": PhoneNumber, 'x_02': slotnum, 'x_03': d.server.adb.device_serial( ),
+                            'x_04': featureCodeInfo, 'x_05': nowTime, 'x_19': 'WXRegister'}
+                    self.repo.PostInformation( saveCate, para )
+
+                    # 入卡槽信息
+                    self.slot.backup( slotnum, str( slotnum ) + '_' + PhoneNumber )  # 卡槽号，手机号
                     return "success"
 
                 for i in range( 1, 6 ):
@@ -314,36 +273,26 @@ class WeiXinRegister:
                         if d( className='android.widget.RadioButton', index=1 ).exists:
                             number = random.randint( 1, 5 )
                             if number == 1:
-                                x_value = d( className='android.widget.RadioButton', index=1 ).info[
-                                    'contentDescription'].replace( '\n\n', '' )
                                 d( className='android.widget.RadioButton', index=1 ).click( )
                                 d( className='android.view.View',
                                    descriptionContains='下一步' ).click( )  # 下一步
                                 z.sleep( 2 )
                             if number == 2:
-                                x_value = d( className='android.widget.RadioButton', index=2 ).info[
-                                    'contentDescription'].replace( '\n\n', '' )
                                 d( className='android.widget.RadioButton', index=2 ).click( )
                                 d( className='android.view.View',
                                    descriptionContains='下一步' ).click( )  # 下一步
                                 z.sleep( 2 )
                             if number == 3:
-                                x_value = d( className='android.widget.RadioButton', index=3 ).info[
-                                    'contentDescription'].replace( '\n\n', '' )
                                 d( className='android.widget.RadioButton', index=3 ).click( )
                                 d( className='android.view.View',
                                    descriptionContains='下一步' ).click( )  # 下一步
                                 z.sleep( 2 )
                             if number == 4:
-                                x_value = d( className='android.widget.RadioButton', index=4 ).info[
-                                    'contentDescription'].replace( '\n\n', '' )
                                 d( className='android.widget.RadioButton', index=4 ).click( )
                                 d( className='android.view.View',
                                    descriptionContains='下一步' ).click( )  # 下一步
                                 z.sleep( 2 )
                             if number == 5:
-                                x_value = d( className='android.widget.RadioButton', index=5 ).info[
-                                    'contentDescription'].replace( '\n\n', '' )
                                 d( className='android.widget.RadioButton', index=5 ).click( )
                                 d( className='android.view.View',
                                    descriptionContains='下一步' ).click( )  # 下一步
@@ -370,36 +319,19 @@ class WeiXinRegister:
                             d.click( 410, 640 )
                         z.sleep( 2 )
 
-                x_04_number = self.repo.GetInformation( information_cate_id, PhoneNumber )
-                x_04 = int( x_04_number[0]['x04'] ) + 1
-                para = {"phoneNumber": PhoneNumber, 'x_04': x_04, 'x_19': 'WXRegister'}
-                self.repo.PostInformation( saveCate, para )
-
             z.sleep( 5 )
-
             if d( text='确认登录' ).exists:
-                para = {"phoneNumber": PhoneNumber, 'x_01': "exist", 'x_03': nowTime, 'x_19': 'WXRegister'}
-                self.repo.PostInformation( saveCate, para )
-                break
+                d( className='android.widget.ImageView', description='返回' ).click( )
+                continue
 
             if d( descriptionContains='验证失败', className='android.view.View' ).exists:
                 d( descriptionContains='关闭页面', className='android.view.View', index=3 ).click( )
-                if x_value == '':
-                    para = {"phoneNumber": PhoneNumber, 'x_01': "exist", 'x_05': '空', 'x_19': 'WXRegister'}
-                else:
-                    para = {"phoneNumber": PhoneNumber, 'x_01': "exist",
-                            "x_key": x_value, 'x_19': 'WXRegister'}
-                self.repo.PostInformation( saveCate, para )
+                continue
 
             if d( descriptionContains='验证通过', className='android.view.View' ).exists:
                 d( descriptionContains='关闭页面', className='android.view.View', index=3 ).click( )
-                if x_value == '':
-                    x_value = 'YSE'
-                para = {"phoneNumber": PhoneNumber, 'x_01': "exist", 'x_05': x_value, 'x_19': 'WXRegister'}
-                self.repo.PostInformation( saveCate, para )
-                featureCodeInfo = z.get_serial( "com.tencent.mm" )
-                self.slot.backup( slotnum, str( slotnum ) + '_' + PhoneNumber + "_" + featureCodeInfo )  # 卡槽号，手机号
-                return "success"
+                continue
+
         return "again"
 
 
@@ -433,8 +365,12 @@ class WeiXinRegister:
                 obj = self.slot.getSlotInfo( slotnum )
                 remark = obj['remark']
                 remarkArr = remark.split( "_" )
-                if len( remarkArr ) == 4:
-                    featureCodeInfo = remarkArr[2] + remarkArr[3]
+                if len( remarkArr ) == 2:
+
+                    phonenumber = remarkArr[1]
+                    information_cate_id = args['repo_information_id']
+                    numbers = self.repo.GetInformation( information_cate_id,phonenumber)
+                    featureCodeInfo = numbers[0]['x04']
                     z.set_serial( "com.tencent.mm", featureCodeInfo )
                 self.slot.restore( slotnum )  # 有time_limit分钟没用过的卡槽情况，切换卡槽
                 d.server.adb.cmd("shell", "am broadcast -a com.zunyun.zime.toast --es msg \"卡槽成功切换为" + slotnum + "号\"").communicate()
@@ -487,13 +423,19 @@ if __name__ == "__main__":
     # z.server.install()
     d.server.adb.cmd("shell", "ime set com.zunyun.qk/.ZImeService").communicate()
     repo = Repo()
-    # repo.RegisterAccount('', 'gemb1225', '13045537833', '109')
-    args = {"repo_information_id": "191"}  # cate_id是仓库号，发中文问题
+    args = {"repo_information_id": "210"}  # cate_id是仓库号，发中文问题
     o.action(d, z, args)
-
-
-
-
+    # serial = d.server.adb.device_serial( )
+    # slot = Slot( serial, 'wechat' )
+    # slot.clear( 1 )('36.149.67.161', None)('PING baidu.com (111.13.101.208)
+    # while True:
+    #     ping = d.server.adb.cmd( "shell", "ping -c 3 baidu.com" ).communicate( )
+    #     ping1 = d.server.adb.cmd( "shell", "curl http://ipecho.net/plain" ).communicate( )
+    #     print( ping )
+    #     print( ping1 )
+    #     if 'icmp_seq' and 'bytes from' and 'time' in ping[0]:
+    #         break
+    #     z.sleep( 2 )
